Add generated truss code and proxy bigbrother checkins to handlers
Add big brother proxy, just forwards big brother requests to truss. This
is nessasary because truss does not support query strings in the
POSTBODY that area is reserved for json. The proxy takes the BODY and
reencodes onto the query string in the URL.

Start of aerospike put and get. Need to configure serverside and
generally figure out how configuration will be instrumented.